requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]

scikit-learn
dill
pyyaml
xgboost
fastapi
uvicorn
ipykernal
apache-airflow
mlflow
dvc
starlette
-e . 